geotiff tests: consolidate predictor cluster (predictor 2 / 3)
Fold six top-level predictor test files into a new
`xrspatial/geotiff/tests/unit/test_predictor.py`, parametrised over
predictor in {2, 3} x dtype x endianness x layout (and multi-sample
for the chunky-band paths).

Files folded:
- test_predictor2_big_endian.py
- test_predictor2_int8.py
- test_predictor3_big_endian.py
- test_predictor3_int_dtype_1933.py
- test_predictor_fp_write_1313.py
- test_predictor_multisample.py

Net file delta: 6 deleted, 1 added => -5 files.

Behaviour preserved:
- predictor=2 round-trips across uint8/int8/uint16/int16/uint32/int32,
  big- and little-endian, stripped and tiled.
- predictor=3 round-trips for float32/float64, big- and little-endian,
  stripped and tiled (read + writer end-to-end).
- libtiff/GDAL multi-byte sample-level differencing interop.
- Multi-sample predictor=2 GPU parity and predictor=3 multi-sample
  TN3 decode (issues #1220, #1247, #1479).
- Predictor=3 + integer SampleFormat validator (#1933).
- normalize_predictor bool|int -> int mapping and rejection branches.
- The opt-in perf gate (predictor=3 within 2x of predictor=2).

Boundary: GPU predictor files
(`test_predictor2_big_endian_gpu_1517.py`,
`test_predictor3_int_dtype_gpu_1933.py`) stay in the GPU cluster
issue #2438. GPU regressions co-located with CPU tests in the old
files move with this consolidation so behaviour-grouped coverage
stays together (matches PR #2451).

Audit: `xrspatial/geotiff/tests/CLUSTER_AUDIT_PREDICTOR.md` maps every
old `file::test` (with parametrise expansion) to the new
`file::test_id`. Deleted on the pre-merge commit before this lands on
main.

Closes #2427.
